+++ src/out_hoi4/ideas/out_ideas.cpp
@@ -35,7 +35,7 @@ void out::OutputMonarchIdeas(const path& output_name,
          continue;
       }
 
-      const auto monarch_itr = characters.find(*monarch_idea_id);
+      const auto monarch_itr = characters.find(static_cast<int>(*monarch_idea_id));
       if (monarch_itr == characters.end())
       {
          continue;
